card throw sound effects. 4 effects picked randomly with slightly randomized pitch. The same effect will never play twice in a row
