+++ Assets/Sound/Card Throw.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7c02883d1da820b47954da0ce5dee189
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
