+++ Assets/_Scripts/Sound.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 34b8ac4ce7608dd498d058ad2769c4e5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
