+++ Assets/_Scripts/Sound/FootStepManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 34f071d45c1b0db4f8e1d348e5a71605
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
